+++ omniva-blocks-builder/src/terminal-selection-block/checkout/options.js
@@ -1,0 +1,17 @@
+/**
+ * External dependencies
+ */
+import { __ } from '@wordpress/i18n';
+
+import { txt } from '../global/text';
+
+export const options = [
+	{
+		label: txt.select_terminal,
+		value: '',
+	},
+	{
+		label: "Testas",
+		value: 'test',
+	},
+];