+++ omniva-woocommerce/assets/blocks/terminal-selection-block/cart/frontend.asset.php
@@ -1,0 +1,1 @@
+<?php return array('dependencies' => array('react', 'wc-blocks-checkout', 'wp-data', 'wp-element', 'wp-i18n'), 'version' => '0781760c9724436c665a');